refactor: move client and server to new directories
