+++ src/server/cmd/subscribe.rs
@@ -1,5 +1,5 @@
-use crate::cmd::{Parse, ParseError};
-use crate::{Command, Connection, Frame, Kv, Shutdown};
+use crate::server::cmd::{Command, Parse, ParseError};
+use crate::{Connection, Frame, Kv, Shutdown};
 
 use bytes::Bytes;
 use std::io;
